+++ src/messages/request.rs
@@ -22,7 +22,7 @@ impl Message for Request {
     const SIZE: MessageLength = MessageLength::Fixed(13);
     const NAME: &'static str = "Request";
 
-    fn read_data<T: Read>(reader: &mut T, _: u32) -> Result<Self, Error> {
+    fn read_data<T: Read>(reader: &mut T, _: usize) -> Result<Self, Error> {
         let index = read_u32(reader)? as usize;
         let begin = read_u32(reader)? as usize;
         let length = read_u32(reader)? as usize;
